+++ windows/src/desktop/kmshell/main/Keyman.System.UpdateStateMachine.pas
@@ -567,7 +567,7 @@ function TUpdateStateMachine.ReadyToInstall: Boolean;
 begin
   if not IsCurrentStateAssigned then
     Exit(False);
-  if (CurrentState.ClassName = 'WaitingRestartState') and not HasKeymanRun then
+  if (CurrentState is WaitingRestartState) and not HasKeymanRun then
     Result := True
   else
     Result := False;
@@ -767,6 +767,7 @@ begin
     // IdleState to wait 'CheckPeriod' before trying again
     TKeymanSentryClient.Client.MessageEvent(Sentry.Client.SENTRY_LEVEL_ERROR,
     'Error Updates not downloaded after 3 attempts');
+    bucStateContext.RemoveCachedFiles;
     ChangeState(IdleState);
   end
   else
@@ -996,17 +997,28 @@ function InstallingState.DoInstallPackage(PackageFileName: String): Boolean;
 var
   FPackage: IKeymanPackageFile2;
 begin
-  Result := True;
-  FPackage := kmcom.Packages.GetPackageFromFile(PackageFileName)
-    as IKeymanPackageFile2;
-  FPackage.Install2(True);
-  // Force overwrites existing package and leaves most settings for it intact
+  Result := False;
+  try
+    FPackage := kmcom.Packages.GetPackageFromFile(PackageFileName)
+      as IKeymanPackageFile2;
+      // Force overwrites existing package and leaves most settings for it intact
+    FPackage.Install2(True);
+    Result := True;
+  except
+    on E:Exception do
+    begin
+      TKeymanSentryClient.ReportHandledException(E,
+        'Failed to install keyboard package');
+    end;
+  end;
+
   FPackage := nil;
-
-  kmcom.Refresh;
-  kmcom.Apply;
-
-  System.SysUtils.DeleteFile(PackageFileName);
+  if Result then
+  begin
+    kmcom.Refresh;
+    kmcom.Apply;
+    System.SysUtils.DeleteFile(PackageFileName);
+  end;
 
 end;
 
@@ -1021,9 +1033,15 @@ begin
   for i := 0 to High(Params.Packages) do
   begin
     PackageFullPath := SavePath + Params.Packages[i].FileName;
+    if not FileExists(PackageFullPath) then
+    begin
+      Continue;
+    end;
+
     if not DoInstallPackage(PackageFullPath) then // I2742
     begin
-      KL.Log('Installing Package failed' + PackageFullPath);
+      TKeymanSentryClient.Client.MessageEvent(Sentry.Client.SENTRY_LEVEL_ERROR,
+      'Installing Package failed"' + PackageFullPath + '"');
     end;
   end;
   Result := True;
